Добавил создание структуры сервисом Github. Изменил вывод флеш сообщений.
