--- lib/clients/http_github_client.rb
@@ -1,24 +1,0 @@
-class GithubClient
-
-  ROOT_ENDPOINT = 'https://api.github.com'
-  ACCESS_TOKEN = ENV['ACCESS_TOKEN']
-
-  def initialize
-    @http_client = setup_http_client
-  end
-
-  def create_gists(params)
-    @http_client.post('gists') do |request|
-      request.headers['Authorization'] = "token #{ACCESS_TOKEN}"
-      request.headers['Content-type'] = 'application/json'
-      request.body = params.to_json
-    end
-  end
-
-  private
-
-  def setup_http_client
-    Faraday.new(url: ROOT_ENDPOINT)
-  end
-
-end
